Trait oak_containers_orchestrator::proto::oak::containers::orchestrator_server::Orchestrator

pub trait Orchestrator: Send + Sync + 'static {
    // Required methods
    fn get_application_config<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetApplicationConfigResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn notify_app_ready<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with OrchestratorServer.

Required Methods

fn get_application_config<'life0, 'async_trait>( &'life0 self, request: Request<()> ) -> Pin<Box<dyn Future<Output = Result<Response<GetApplicationConfigResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Exposes the previously loaded trusted application config to the application, which may choose to retrieve it.

fn notify_app_ready<'life0, 'async_trait>( &'life0 self, request: Request<()> ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Notifies the orchestrator that the trusted app is ready to serve requests and listening on the pre-arranged port (8080).